Обновить сущности и DTO для поддержки Long идентификаторов; добавить маппинг между Task и TaskDto; изменить контроллеры и сервисы для работы с новыми типами данных; улучшить структуру проекта.

+++ src/main/java/com/arthwrntzch/SpringBootToDoList/controller/UserController.java
@@ -31,7 +31,7 @@ public class UserController {
     public List<User> getAllUsers() { return userService.getAllUsers(); }
 
     @GetMapping("/{id}")
-    public ResponseEntity<User> getUserById(@PathVariable int id) {
+    public ResponseEntity<User> getUserById(@PathVariable Long id) {
         User user = userService.getUserById(id);
         return user != null ? ResponseEntity.ok(user)
                 : ResponseEntity.notFound().build();
