+++ src/main/java/com/arthwrntzch/SpringBootToDoList/repository/UserRepository.java
@@ -1,10 +1,7 @@
 package com.arthwrntzch.SpringBootToDoList.repository;
 
+import com.arthwrntzch.SpringBootToDoList.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-
-import com.arthwrntzch.SpringBootToDoList.entity.User;
-
-public interface UserRepository extends JpaRepository<User, Integer>{
-    
+public interface UserRepository extends JpaRepository<User, Long> {
 }
